pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tourney-threads"
version = "0.1.0"
description = "Create Discord threads for Challonge matchups (supports Swiss/group stages)."
readme = "README.md"
requires-python = ">=3.9"
authors = [{ name = "Your Team" }]
license = { text = "MIT" }
dependencies = [
  "discord.py>=2.4.0",
  "aiohttp>=3.9.5",
  "PyYAML>=6.0.1"
]

[project.scripts]
tourney-threads = "tourney_threads.app:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
